# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Ocean Weligama API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /v1/admin/services:
    get:
      operationId: adminListServices
      tags: [admin-services]
      summary: List all services
      responses:
        "200":
          description: All services
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Service"
    post:
      operationId: adminCreateService
      tags: [admin-services]
      summary: Create service
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ServiceInput"
      responses:
        "201":
          description: Service created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Service"
  /v1/admin/services/{id}:
    patch:
      operationId: adminUpdateService
      tags: [admin-services]
      summary: Update service
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ServiceUpdate"
      responses:
        "200":
          description: Updated service
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Service"
    delete:
      operationId: adminDeleteService
      tags: [admin-services]
      summary: Delete service
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Service deleted
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean

  # ADMIN REVIEWS
...
  /v1/admin/reviews/{id}:
    patch:
      operationId: adminUpdateReview
      tags: [admin-reviews]
      summary: Approve, reply to, or feature a review
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReviewUpdate"
      responses:
        "200":
          description: Updated review
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Review"
    delete:
      operationId: adminDeleteReview
      tags: [admin-reviews]
      summary: Delete a review permanently
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Deleted

  # ADMIN GALLERY
...
  /v1/admin/messages:
    get:
      operationId: adminListMessages
      tags: [admin-messages]
      summary: List contact messages
      parameters:
        - name: isRead
          in: query
          schema:
            type: boolean
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: Messages
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ContactMessage"
  /v1/admin/messages/{id}:
    patch:
      operationId: adminUpdateMessage
      tags: [admin-messages]
      summary: Mark message read or replied
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MessageUpdate"
      responses:
        "200":
          description: Updated message
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ContactMessage"
    delete:
      operationId: adminDeleteMessage
      tags: [admin-messages]
      summary: Delete a message permanently
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Deleted

  # ADMIN PRICING
  /v1/admin/pricing:
    get:
      operationId: adminListPricingRules
      tags: [admin-pricing]
      summary: List all pricing rules
      responses:
        "200":
          description: Pricing rules
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PricingRule"
    post:
      operationId: adminCreatePricingRule
      tags: [admin-pricing]
      summary: Create pricing rule
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PricingRuleInput"
      responses:
        "201":
          description: Created pricing rule
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PricingRule"
  /v1/admin/pricing/{id}:
    patch:
      operationId: adminUpdatePricingRule
      tags: [admin-pricing]
      summary: Update pricing rule
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PricingRuleUpdate"
      responses:
        "200":
          description: Updated pricing rule
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PricingRule"
    delete:
      operationId: adminDeletePricingRule
      tags: [admin-pricing]
      summary: Delete pricing rule
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Deleted

  # ADMIN AVAILABILITY
  /v1/admin/availability/{roomId}:
    get:
      operationId: adminGetRoomAvailability
      tags: [admin-availability]
      summary: Get room availability calendar
      parameters:
        - name: roomId
          in: path
          required: true
          schema:
            type: string
        - name: from
          in: query
          schema:
            type: string
            format: date
        - name: to
          in: query
          schema:
            type: string
            format: date
      responses:
        "200":
          description: Availability
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AvailabilityDay"
  /v1/admin/availability/block:
    post:
      operationId: adminBlockDates
      tags: [admin-availability]
      summary: Block dates for a room
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BlockDatesInput"
      responses:
        "201":
          description: Dates blocked
...
  /v1/admin/availability/{id}:
    delete:
      operationId: adminUnblockDate
      tags: [admin-availability]
      summary: Unblock a date
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Unblocked

  # PUBLIC BLOGS
...
components:
  schemas:
...
    Service:
      type: object
      required: [id, slug, basePrice, isActive, name, type]
      properties:
        id:
          type: string
        slug:
          type: string
        imageUrl:
          type: ["string", "null"]
        highlights:
          type: array
          items:
            type: string
        type:
          type: string
          enum: [main, optional]
        category:
          type: ["string", "null"]
        basePrice:
          type: string
        isActive:
          type: boolean
        isBookable:
          type: boolean
        sortOrder:
          type: integer
        name:
          type: string
        description:
          type: ["string", "null"]
    ServiceInput:
      type: object
      required: [slug, basePrice, name, type]
      properties:
        slug:
          type: string
        imageUrl:
          type: ["string", "null"]
        highlights:
          type: array
          items:
            type: string
        type:
          type: string
          enum: [main, optional]
        category:
          type: ["string", "null"]
        basePrice:
          type: string
        isActive:
          type: boolean
        isBookable:
          type: boolean
        name:
          type: string
        description:
          type: ["string", "null"]
    ServiceUpdate:
      type: object
      properties:
        slug:
          type: string
        imageUrl:
          type: ["string", "null"]
        highlights:
          type: array
          items:
            type: string
        type:
          type: string
          enum: [main, optional]
        category:
          type: ["string", "null"]
        basePrice:
          type: string
        isActive:
          type: boolean
        isBookable:
          type: boolean
        name:
          type: string
        description:
          type: ["string", "null"]
        sortOrder:
          type: integer
...
    Review:
      type: object
      required: [id, guestName, ratingOverall, reviewText, isApproved, createdAt]
      properties:
        id:
          type: string
        bookingId:
          type: ["string", "null"]
        guestName:
          type: string
        guestCountry:
          type: ["string", "null"]
        guestAvatarUrl:
          type: ["string", "null"]
        ratingOverall:
          type: integer
        title:
          type: ["string", "null"]
        ratingCleanliness:
          type: ["integer", "null"]
        ratingLocation:
          type: ["integer", "null"]
        ratingValue:
          type: ["integer", "null"]
        reviewText:
          type: string
        ownerReply:
          type: ["string", "null"]
        isApproved:
          type: boolean
        isFeatured:
          type: boolean
        source:
          type: string
        createdAt:
          type: string
...
    ReviewUpdate:
      type: object
      properties:
        isApproved:
          type: boolean
        isFeatured:
          type: boolean
        ownerReply:
          type: ["string", "null"]
...
    AvailabilityDay:
      type: object
      required: [date, isBlocked]
      properties:
        date:
          type: string
        isBlocked:
          type: boolean
        blockReason:
          type: ["string", "null"]
...
    ContactMessage:
      type: object
      required: [id, fullName, email, message, isRead, isReplied, createdAt]
      properties:
        id:
          type: string
        fullName:
          type: string
        email:
          type: string
        phone:
          type: ["string", "null"]
        subject:
          type: ["string", "null"]
        message:
          type: string
        countryCode:
          type: ["string", "null"]
        isRead:
          type: boolean
        isReplied:
          type: boolean
        repliedAt:
          type: ["string", "null"]
        createdAt:
          type: string
    MessageUpdate:
      type: object
      properties:
        isRead:
          type: boolean
        isReplied:
          type: boolean
...
    PricingRule:
      type: object
      required: [id, name, dateFrom, dateTo, pricePerNight, isActive]
      properties:
        id:
          type: string
        roomId:
          type: ["string", "null"]
        name:
          type: string
        dateFrom:
          type: string
        dateTo:
          type: string
        pricePerNight:
          type: string
        minNights:
          type: string
        isActive:
          type: boolean
        priority:
          type: string
    PricingRuleInput:
      type: object
      required: [name, dateFrom, dateTo, pricePerNight]
      properties:
        roomId:
          type: ["string", "null"]
        name:
          type: string
        dateFrom:
          type: string
        dateTo:
          type: string
        pricePerNight:
          type: string
        minNights:
          type: string
        isActive:
          type: boolean
        priority:
          type: string
    PricingRuleUpdate:
      type: object
      properties:
        name:
          type: string
        dateFrom:
          type: string
        dateTo:
          type: string
        pricePerNight:
          type: string
        isActive:
          type: boolean
    BlockDatesInput:
      type: object
      required: [roomId, dates, blockReason]
      properties:
        roomId:
          type: string
        dates:
          type: array
          items:
            type: string
        blockReason:
          type: string